+++ app/Http/Controllers/RestaurantController.php
@@ -1,7 +1,15 @@
 <?php
+/**
+ * File name: RestaurantController.php
+ * Last modified: 2020.04.30 at 08:21:08
+ * Author: SmarterVision - https://codecanyon.net/user/smartervision
+ * Copyright (c) 2020
+ *
+ */
 
 namespace App\Http\Controllers;
 
+use App\Criteria\Restaurants\RestaurantsOfUserCriteria;
 use App\Criteria\Users\DriversCriteria;
 use App\Criteria\Users\ManagersCriteria;
 use App\DataTables\RestaurantDataTable;
@@ -9,6 +17,7 @@ use App\Events\RestaurantChangedEvent;
 use App\Http\Requests\CreateRestaurantRequest;
 use App\Http\Requests\UpdateRestaurantRequest;
 use App\Repositories\CustomFieldRepository;
+use App\Repositories\CuisineRepository;
 use App\Repositories\RestaurantRepository;
 use App\Repositories\UploadRepository;
 use App\Repositories\UserRepository;
@@ -36,14 +45,19 @@ class RestaurantController extends Controller
      * @var UserRepository
      */
     private $userRepository;
+    /**
+     * @var CuisineRepository
+     */
+    private $cuisineRepository;
 
-    public function __construct(RestaurantRepository $restaurantRepo, CustomFieldRepository $customFieldRepo, UploadRepository $uploadRepo, UserRepository $userRepo)
+    public function __construct(RestaurantRepository $restaurantRepo, CustomFieldRepository $customFieldRepo, UploadRepository $uploadRepo, UserRepository $userRepo, CuisineRepository $cuisineRepository)
     {
         parent::__construct();
         $this->restaurantRepository = $restaurantRepo;
         $this->customFieldRepository = $customFieldRepo;
         $this->uploadRepository = $uploadRepo;
         $this->userRepository = $userRepo;
+        $this->cuisineRepository = $cuisineRepository;
     }
 
     /**
@@ -67,14 +81,16 @@ class RestaurantController extends Controller
 
         $user = $this->userRepository->getByCriteria(new ManagersCriteria())->pluck('name', 'id');
         $drivers = $this->userRepository->getByCriteria(new DriversCriteria())->pluck('name', 'id');
+        $cuisine = $this->cuisineRepository->pluck('name', 'id');
         $usersSelected = [];
         $driversSelected = [];
+        $cuisinesSelected = [];
         $hasCustomField = in_array($this->restaurantRepository->model(), setting('custom_field_models', []));
         if ($hasCustomField) {
             $customFields = $this->customFieldRepository->findByField('custom_field_model', $this->restaurantRepository->model());
             $html = generateCustomField($customFields);
         }
-        return view('restaurants.create')->with("customFields", isset($html) ? $html : false)->with("user", $user)->with("drivers", $drivers)->with("usersSelected", $usersSelected)->with("driversSelected", $driversSelected);
+        return view('restaurants.create')->with("customFields", isset($html) ? $html : false)->with("user", $user)->with("drivers", $drivers)->with("usersSelected", $usersSelected)->with("driversSelected", $driversSelected)->with('cuisine', $cuisine)->with('cuisinesSelected', $cuisinesSelected);
     }
 
     /**
@@ -89,7 +105,6 @@ class RestaurantController extends Controller
         $input = $request->all();
         if (auth()->user()->hasRole('manager')) {
             $input['users'] = [auth()->id()];
-            unset($input['admin_commission']);
         }
         $customFields = $this->customFieldRepository->findByField('custom_field_model', $this->restaurantRepository->model());
         try {
@@ -116,9 +131,11 @@ class RestaurantController extends Controller
      * @param int $id
      *
      * @return Response
+     * @throws \Prettus\Repository\Exceptions\RepositoryException
      */
     public function show($id)
     {
+        $this->restaurantRepository->pushCriteria(new RestaurantsOfUserCriteria(auth()->id()));
         $restaurant = $this->restaurantRepository->findWithoutFail($id);
 
         if (empty($restaurant)) {
@@ -136,23 +153,27 @@ class RestaurantController extends Controller
      * @param int $id
      *
      * @return Response
+     * @throws \Prettus\Repository\Exceptions\RepositoryException
      */
     public function edit($id)
     {
+        $this->restaurantRepository->pushCriteria(new RestaurantsOfUserCriteria(auth()->id()));
         $restaurant = $this->restaurantRepository->findWithoutFail($id);
+
+        if (empty($restaurant)) {
+            Flash::error(__('lang.not_found', ['operator' => __('lang.restaurant')]));
+            return redirect(route('restaurants.index'));
+        }
 
         $user = $this->userRepository->getByCriteria(new ManagersCriteria())->pluck('name', 'id');
         $drivers = $this->userRepository->getByCriteria(new DriversCriteria())->pluck('name', 'id');
+        $cuisine = $this->cuisineRepository->pluck('name', 'id');
 
 
         $usersSelected = $restaurant->users()->pluck('users.id')->toArray();
         $driversSelected = $restaurant->drivers()->pluck('users.id')->toArray();
+        $cuisinesSelected = $restaurant->cuisines()->pluck('cuisines.id')->toArray();
 
-        if (empty($restaurant)) {
-            Flash::error(__('lang.not_found', ['operator' => __('lang.restaurant')]));
-
-            return redirect(route('restaurants.index'));
-        }
         $customFieldsValues = $restaurant->customFieldsValues()->with('customField')->get();
         $customFields = $this->customFieldRepository->findByField('custom_field_model', $this->restaurantRepository->model());
         $hasCustomField = in_array($this->restaurantRepository->model(), setting('custom_field_models', []));
@@ -160,7 +181,7 @@ class RestaurantController extends Controller
             $html = generateCustomField($customFields, $customFieldsValues);
         }
 
-        return view('restaurants.edit')->with('restaurant', $restaurant)->with("customFields", isset($html) ? $html : false)->with("user", $user)->with("drivers", $drivers)->with("usersSelected", $usersSelected)->with("driversSelected", $driversSelected);
+        return view('restaurants.edit')->with('restaurant', $restaurant)->with("customFields", isset($html) ? $html : false)->with("user", $user)->with("drivers", $drivers)->with("usersSelected", $usersSelected)->with("driversSelected", $driversSelected)->with('cuisine', $cuisine)->with('cuisinesSelected', $cuisinesSelected);
     }
 
     /**
@@ -170,9 +191,11 @@ class RestaurantController extends Controller
      * @param UpdateRestaurantRequest $request
      *
      * @return Response
+     * @throws \Prettus\Repository\Exceptions\RepositoryException
      */
     public function update($id, UpdateRestaurantRequest $request)
     {
+        $this->restaurantRepository->pushCriteria(new RestaurantsOfUserCriteria(auth()->id()));
         $restaurant = $this->restaurantRepository->findWithoutFail($id);
 
         if (empty($restaurant)) {
@@ -182,9 +205,8 @@ class RestaurantController extends Controller
         $input = $request->all();
         $customFields = $this->customFieldRepository->findByField('custom_field_model', $this->restaurantRepository->model());
         try {
+
             $restaurant = $this->restaurantRepository->update($input, $id);
-            $input['users'] = isset($input['users']) ? $input['users'] : [];
-            $input['drivers'] = isset($input['drivers']) ? $input['drivers'] : [];
             if (isset($input['image']) && $input['image']) {
                 $cacheUpload = $this->uploadRepository->getByUuid($input['image']);
                 $mediaItem = $cacheUpload->getMedia('image')->first();
@@ -210,21 +232,26 @@ class RestaurantController extends Controller
      * @param int $id
      *
      * @return Response
+     * @throws \Prettus\Repository\Exceptions\RepositoryException
      */
     public function destroy($id)
     {
-        $restaurant = $this->restaurantRepository->findWithoutFail($id);
+        if (!env('APP_DEMO', false)) {
+            $this->restaurantRepository->pushCriteria(new RestaurantsOfUserCriteria(auth()->id()));
+            $restaurant = $this->restaurantRepository->findWithoutFail($id);
 
-        if (empty($restaurant)) {
-            Flash::error('Restaurant not found');
+            if (empty($restaurant)) {
+                Flash::error('Restaurant not found');
 
-            return redirect(route('restaurants.index'));
+                return redirect(route('restaurants.index'));
+            }
+
+            $this->restaurantRepository->delete($id);
+
+            Flash::success(__('lang.deleted_successfully', ['operator' => __('lang.restaurant')]));
+        } else {
+            Flash::warning('This is only demo app you can\'t change this section ');
         }
-
-        $this->restaurantRepository->delete($id);
-
-        Flash::success(__('lang.deleted_successfully', ['operator' => __('lang.restaurant')]));
-
         return redirect(route('restaurants.index'));
     }
